requirements_hf.txt
# Hugging Face Spaces Requirements
# Optimized for deployment on HF infrastructure

# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async support
aiohttp>=3.10.0
aiohttp-socks>=0.8.0

# Exchange APIs
ccxt>=4.0.0

# AWS DynamoDB
boto3>=1.34.0

# Google Sheets
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2
google-auth-oauthlib

# Configuration
python-dotenv>=1.0.0

# Gradio Web Interface
gradio>=4.0.0
pandas>=2.0.0

# HTTP requests for Pushover
requests>=2.31.0

# JWT Authentication
PyJWT>=2.8.0
